bots/rhizoo-alpha-bot/services/dashboard/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
redis[hiredis]>=5.0.0
websockets>=12.0
python-dotenv>=1.0.0
